# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kotemon Jastip API
  description: |
    API for Kotemon Jastip - A personal shopping service from Japan to Indonesia.

    ## Authentication

    This API uses two authentication methods:

    1. **Cookie-based** (`cookieAuth`): For regular user sessions after Google OAuth
       - Session cookie name: `jastip_session`
       - Used by: `/api/auth/*` endpoints (except login/init)

    2. **Bearer Token** (`bearerAuth`): For admin access
       - Header: `Authorization: Bearer <admin_password>`
       - Used by: All `/api/admin/*` endpoints

    ## Pricing Model

    The API uses a three-field pricing model:

    - **basePriceYen** (integer): Original price in Japanese Yen (admin enters)
    - **basePriceRp** (integer): Calculated base price in IDR = `basePriceYen × exchangeRate`
    - **sellingPriceRp** (integer): Final selling price in IDR (manual entry)

    **Profit calculation:** `profit = sellingPriceRp - basePriceRp`

    The `basePriceRp` is automatically calculated when creating or updating items based on the current
    `exchangeRate` stored in settings. When `basePriceYen` is updated via PATCH, `basePriceRp` is
    recalculated using the current exchange rate.

    ## Environment Variables

    Required environment variables:
    - `JWT_SECRET` - Secret for signing session tokens
    - `GOOGLE_CLIENT_SECRET` - Google OAuth client secret
    - `DB` - D1 database binding
    - `PHOTOS_BUCKET` - R2 bucket for photo storage
...
  contact:
...
servers:
...
paths:
...
  /public/config:
    get:
      summary: Get public jastip configuration
      description: Retrieves current jastip status, countdown, and baggage quota information
      tags:
        - Public
      operationId: getPublicConfig
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PublicConfig'
        '500':
          $ref: '#/components/responses/InternalError'
  /public/items:
    get:
      summary: List available items
      description: Retrieves all published and available items for public viewing
      tags:
        - Public
      operationId: listPublicItems
      parameters:
        - name: search
          in: query
          description: Search term to filter items by name
          schema:
            type: string
            example: "pokemon"
        - name: limit
          in: query
          description: Maximum number of items to return (max 100)
          schema:
            type: integer
            default: 20
            maximum: 100
            example: 20
        - name: offset
          in: query
          description: Number of items to skip for pagination
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PublicItem'
                      total:
                        type: integer
                        description: Total number of items returned
                        example: 10
        '500':
          $ref: '#/components/responses/InternalError'
  /public/items/{id}:
    get:
      summary: Get single item details
      description: Retrieves detailed information about a specific item
      tags:
        - Public
      operationId: getPublicItem
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the item
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PublicItem'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /public/items/{id}/view:
    post:
      summary: Increment item view count
      description: Increments the view counter for an item and returns the new count
      tags:
        - Public
      operationId: incrementItemView
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the item
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: View count incremented successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  viewCount:
                    type: integer
                    description: New view count after increment
                    example: 42
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jastip_session
      description: Session cookie for authenticated users (7-day expiry)
    bearerAuth:
      type: http
      scheme: bearer
      description: Admin password used as bearer token
  schemas:
...
    ItemCategory:
      type: string
      enum: [snack, skincare, makeup, stationery, gift, beverage, accessories]
      description: Item category for filtering and organization
      example: "snack"
    InfoNoteType:
      type: string
      enum: [amber, purple, blue, red]
      description: Info note badge color type
      example: "amber"
    InfoNote:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/InfoNoteType'
        text:
          type: string
          description: Info note text content
          example: "Limited stock available"
      required:
        - type
        - text

    # Public schemas
    PublicConfig:
      type: object
      properties:
        jastipStatus:
          type: string
          enum: [open, closed]
          description: Current jastip status
          example: "open"
        countdownDays:
          type: integer
          nullable: true
          description: Days until jastip closes (null if closed or no close date)
          example: 5
        remainingQuotaKg:
          type: number
          format: float
          description: Remaining baggage quota in kilograms (1 decimal precision)
          example: 15.5
        totalQuotaKg:
          type: number
          format: float
          description: Total baggage quota in kilograms
          example: 20.0
        estimatedArrivalDate:
          type: string
          nullable: true
          description: Estimated arrival date in Indonesia
          example: "6-10 May 2025"
    PublicItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Pokemon Center Pikachu Plush"
        description:
          type: string
          nullable: true
          example: "Limited edition Pikachu plush from Pokemon Center Tokyo"
        photos:
          type: array
          items:
            type: string
            format: uri
          example: ["/api/public/photos/uploads/abc123.jpg"]
        basePriceRp:
          type: integer
          description: Calculated base price in IDR (basePriceYen × exchangeRate)
          example: 271250
        sellingPriceRp:
          type: integer
          description: Selling price in Indonesian Rupiah (IDR)
          example: 350000
        weightGrams:
          type: integer
          description: Item weight in grams
          example: 250
        withoutBoxNote:
          type: boolean
          description: Show "tanpa box" (without box) info
          example: false
        isLimitedEdition:
          type: boolean
          description: Show "Limited Edition" badge
          example: true
        isPreorder:
          type: boolean
          description: Show "Pre-order" info
          example: false
        isFragile:
          type: boolean
          description: Show "Fragile" warning
          example: true
        category:
          $ref: '#/components/schemas/ItemCategory'
        infoNotes:
          type: array
          items:
            $ref: '#/components/schemas/InfoNote'
          description: Array of info notes with colored badges
          example: [{"type": "amber", "text": "Limited stock available"}, {"type": "blue", "text": "Official merchandise"}]
        availableSlots:
          type: integer
          description: Number of remaining order slots (maxOrders - currentOrders)
          example: 5
        badge:
          type: string
          enum: [available, low_stock, full, new]
          description: |
            Item badge:
            - `available`: Normal availability
            - `low_stock`: 1-2 slots remaining
            - `full`: No slots available
            - `new`: Created within last 3 days
          example: "new"
        viewCount:
          type: integer
          description: Number of times the item has been viewed
          example: 42

    # Auth schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request"
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Item not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"